+++ src/hyprland_config/_hyprlang/__init__.py
@@ -20,7 +20,7 @@ from hyprland_config._hyprlang._parser import (
     parse_file,
     parse_string,
 )
-from hyprland_config._hyprlang._serializer import serialize_hyprlang
+from hyprland_config._hyprlang._serializer import render_rule_hyprlang, serialize_hyprlang
 from hyprland_config._hyprlang._source import SourceCycleError
 
 __all__ = [
@@ -31,5 +31,6 @@ __all__ = [
     "parse_bind_line",
     "parse_file",
     "parse_string",
+    "render_rule_hyprlang",
     "serialize_hyprlang",
 ]